version.py, link_module.py, and panic_test.py were also left out of 2.0 for some reason....
git-svn-id: https://simian.googlecode.com/svn/trunk@70 368532df-a33e-f994-cbfd-937b2e4ba68d

+++ src/simian/util/link_module.sh
@@ -2,8 +2,9 @@
 #
 # Copyright 2011 Google Inc. All Rights Reserved.
 
-GAE_BUNDLE=gae_bundle/
+set -e
 
+GAE_BUNDLE=gae_bundle/
 
 function find_module() {
   python <<EOF
@@ -15,23 +16,36 @@ except ImportError:
 EOF
 }
 
-function find_egg() {
-  egg=$(find ${module}-*.egg -type f 2>/dev/null)
+function find_egg_file() {
+  egg=$(find . -type f -maxdepth 1 -name ${module}-*.egg 2>/dev/null)
   if [[ ! -z "${egg}" ]]; then
     echo "${egg}"
   fi
 }
 
+function find_egg_dir() {
+  egg=$(find . -type d -name ${module}-*.egg 2>/dev/null)
+  if [[ ! -z "${egg}" ]]; then
+    find "${egg}" -type d -maxdepth 1 -mindepth 1 \! -name EGG-INFO
+  fi
+}
+
 function link_module() {
   local module="$1"
-  
-  local egg=$(find_egg ${module})
+
+  local egg=$(find_egg_file ${module})
   if [[ ! -z "${egg}" ]]; then
     unzip -o "${egg}" -d "${GAE_BUNDLE}" > /dev/null
     rm -rf "${GAE_BUNDLE}/EGG-INFO"
     return
   fi
-  
+
+  local egg=$(find_egg_dir ${module})
+  if [[ ! -z "${egg}" ]]; then
+    cp -fR "${egg}" "${GAE_BUNDLE}"
+    return
+  fi
+
   local path=$(find_module ${module})
   if [[ ! -z "${path}" ]]; then
     rm -f "${GAE_BUNDLE}/${module}"
@@ -40,6 +54,7 @@ function link_module() {
   fi
 
   echo "ERROR: path not found for ${module}. symlink creation failure."
+  exit 1
 }
 
 link_module "$1"
